Excluding the "RPY" extension.
I noticed that not much use scripts with the extension "RPY" and I'm doing a package from that to support the extension "RPY" used in Ren'Py engine.
In this way the files ".rpy" may be identified as Ren'Py files automatically. Is up to you to delete or not, but I thank you now!
Translated by Google Translate.